+++ packages/rolldown/tests/fixtures/builtin-plugin/dynamic-import-vars/_config.ts
@@ -1,4 +1,7 @@
-import { dynamicImportVarsPlugin, globImportPlugin } from 'rolldown'
+import {
+  dynamicImportVarsPlugin,
+  globImportPlugin,
+} from 'rolldown/experimental'
 import { defineTest } from '@tests'
 import path from 'path'
