+++ packages/rolldown/tests/fixtures/builtin-plugin/glob-import/basic/_config.ts
@@ -1,4 +1,4 @@
-import { globImportPlugin } from 'rolldown'
+import { globImportPlugin } from 'rolldown/experimental'
 import { defineTest } from '@tests'
 import * as path from 'path'
