+++ packages/rolldown/tests/fixtures/builtin-plugin/glob-import/eager/_config.ts
@@ -1,4 +1,4 @@
-import { globImportPlugin } from 'rolldown'
+import { globImportPlugin } from 'rolldown/experimental'
 import { defineTest } from '@tests'
 
 export default defineTest({
