+++ packages/rolldown/tests/fixtures/builtin-plugin/wasm/_config.ts
@@ -1,4 +1,4 @@
-import { wasmPlugin } from '../../../../'
+import { wasmPlugin } from 'rolldown/experimental'
 import { defineTest } from '@tests'
 
 export default defineTest({
